requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
motor==3.3.2
pymongo==4.6.2
sentence-transformers==2.6.1
pdfminer.six==20231228
httpx==0.27.0
python-dotenv==1.0.1
pydantic[email]==2.6.4
pydantic-settings==2.2.1
slowapi==0.1.9
numpy==1.26.4
aiofiles==23.2.1
pytest==8.1.1
pytest-asyncio==0.23.6
flake8==7.0.0
groq==0.9.0
